requirements.txt
yolov5
requests
BeautifulSoup4
pillow
numpy